requirements.txt
pymongo==4.3.3
PyYAML==6.0
Flask==2.3.2
python-benedict==0.31.0

